Anomaly RM#112433:ManufOrder: Improve form loading performance by extracting lightweight on-load services
