+++ axelor-production/src/main/java/com/axelor/apps/production/module/ProductionModule.java
@@ -228,6 +228,8 @@ import com.axelor.apps.production.service.machine.MachineService;
 import com.axelor.apps.production.service.machine.MachineServiceImpl;
 import com.axelor.apps.production.service.manuforder.ManufOrderCheckStockMoveLineService;
 import com.axelor.apps.production.service.manuforder.ManufOrderCheckStockMoveLineServiceImpl;
+import com.axelor.apps.production.service.manuforder.ManufOrderComputeService;
+import com.axelor.apps.production.service.manuforder.ManufOrderComputeServiceImpl;
 import com.axelor.apps.production.service.manuforder.ManufOrderCreateBarcodeService;
 import com.axelor.apps.production.service.manuforder.ManufOrderCreateBarcodeServiceImpl;
 import com.axelor.apps.production.service.manuforder.ManufOrderCreatePurchaseOrderService;
@@ -258,6 +260,8 @@ import com.axelor.apps.production.service.manuforder.ManufOrderService;
 import com.axelor.apps.production.service.manuforder.ManufOrderServiceImpl;
 import com.axelor.apps.production.service.manuforder.ManufOrderSetStockMoveLineService;
 import com.axelor.apps.production.service.manuforder.ManufOrderSetStockMoveLineServiceImpl;
+import com.axelor.apps.production.service.manuforder.ManufOrderStockLocationQueryService;
+import com.axelor.apps.production.service.manuforder.ManufOrderStockLocationQueryServiceImpl;
 import com.axelor.apps.production.service.manuforder.ManufOrderStockMoveService;
 import com.axelor.apps.production.service.manuforder.ManufOrderStockMoveServiceImpl;
 import com.axelor.apps.production.service.manuforder.ManufOrderTrackingNumberService;
@@ -549,5 +553,8 @@ public class ProductionModule extends AxelorModule {
     bind(ManufOrderMessageService.class).to(ManufOrderMessageServiceImpl.class);
     bind(ManufOrderQueryService.class).to(ManufOrderQueryServiceImpl.class);
     bind(ControlEntryServiceImpl.class).to(ControlEntryProductionServiceImpl.class);
+    bind(ManufOrderComputeService.class).to(ManufOrderComputeServiceImpl.class);
+    bind(ManufOrderStockLocationQueryService.class)
+        .to(ManufOrderStockLocationQueryServiceImpl.class);
   }
 }
