+++ axelor-production/src/main/java/com/axelor/apps/production/service/manuforder/ManufOrderService.java
@@ -39,15 +39,6 @@ public interface ManufOrderService {
   public static int DEFAULT_PRIORITY_INTERVAL = 10;
   public static boolean IS_TO_INVOICE = false;
 
-  /**
-   * This method check if operation orders regardless of manufOrder. If manufOrder is outsourced,
-   * the method will return false as they are outsourced because of manufOrder.
-   *
-   * @param manufOrder
-   * @return true if lines are outsourced regardless of manufOrder, else false.
-   */
-  boolean areLinesOutsourced(ManufOrder manufOrder);
-
   void setOperationOrdersOutsourcing(ManufOrder manufOrder);
 
   public interface ManufOrderOriginType {}
@@ -168,25 +159,6 @@ public interface ManufOrderService {
       List<ProdProduct> prodProductList, List<StockMoveLine> stockMoveLineList)
       throws AxelorException;
 
-  /**
-   * Create a query to find product's consume and missing qty of a specific/all company and a
-   * specific/all stock location in a Manuf Order
-   *
-   * @param productId, companyId and stockLocationId
-   * @return the query.
-   */
-  public String getConsumeAndMissingQtyForAProduct(
-      Long productId, Long companyId, Long stockLocationId);
-
-  /**
-   * Create a query to find product's building qty of a specific/all company and a specific/all
-   * stock location in a Manuf Order
-   *
-   * @param productId, companyId and stockLocationId
-   * @return the query.
-   */
-  public String getBuildingQtyForAProduct(Long productId, Long companyId, Long stockLocationId);
-
   public List<ManufOrder> generateAllSubManufOrder(List<Product> productList, ManufOrder manufOrder)
       throws AxelorException;
 
@@ -213,8 +185,6 @@ public interface ManufOrderService {
 
   List<ManufOrder> getChildrenManufOrder(ManufOrder manufOrder);
 
-  public BigDecimal computeProducibleQty(ManufOrder manufOrder) throws AxelorException;
-
   /**
    * Method that will update planned dates of manuf order. Unlike the other methods, this will not
    * reset planned dates of the operation orders of the manuf order. This method must be called when
@@ -225,7 +195,4 @@ public interface ManufOrderService {
   public void updatePlannedDates(ManufOrder manufOrder);
 
   void checkApplicableManufOrder(ManufOrder manufOrder) throws AxelorException;
-
-  public Map<Product, BigDecimal> getMissingComponents(ManufOrder manufOrder)
-      throws AxelorException;
 }
