+++ axelor-production/src/main/java/com/axelor/apps/production/service/manuforder/ManufOrderStockLocationQueryService.java
@@ -1,0 +1,40 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2005-2026 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package com.axelor.apps.production.service.manuforder;
+
+public interface ManufOrderStockLocationQueryService {
+
+  /**
+   * Create a query to find product's consume and missing qty of a specific/all company and a
+   * specific/all stock location in a Manuf Order
+   *
+   * @param productId, companyId and stockLocationId
+   * @return the query.
+   */
+  String getConsumeAndMissingQtyForAProduct(Long productId, Long companyId, Long stockLocationId);
+
+  /**
+   * Create a query to find product's building qty of a specific/all company and a specific/all
+   * stock location in a Manuf Order
+   *
+   * @param productId, companyId and stockLocationId
+   * @return the query.
+   */
+  String getBuildingQtyForAProduct(Long productId, Long companyId, Long stockLocationId);
+}
